Move RedirectStdIn
